added hasura connection via apollo/gql
