+++ src/gql/index.ts
@@ -1,5 +1,5 @@
 import { ApolloClient, InMemoryCache } from '@apollo/client'
-export * from './queries/getUserByName'
+export * from './queries/getUserRoleByName'
 
 const URI = 'https://square-basilisk-55.hasura.app/v1/graphql'
 const ADMIN_SECRET =
